start of refactoring + tests
basic new ppu structure

numerous new tests to the parts that weren't refactored
